Use long long (int64) for cvxcore I/J indices to satisfy PTNobel review
- ProblemData.hpp: TensorI/TensorJ changed from vector<int> to vector<long long>
- ProblemData.hpp: getI/getJ return long long* instead of int*
- cvxcore.cpp: add_matrix_to_vectors takes vector<long long> for I/J
- cvxcore.i: ARGOUT typemap updated from int* to long long* for getI/getJ

This ensures the row/column indices are natively 64-bit throughout the
C++ boundary and arrive in Python as np.int64 arrays, preventing overflow
for large problems and removing MOSEK 11 type-cast warnings.

+++ cvxpy/cvxcore/src/ProblemData.hpp
@@ -25,8 +25,8 @@ public:
   /* COO sparse matrix representation. V stores the data, I the row indices
    * and J the column indices. */
   std::map<int, std::vector<std::vector<double> > > TensorV;
-  std::map<int, std::vector<std::vector<int> > > TensorI;
-  std::map<int, std::vector<std::vector<int> > > TensorJ;
+  std::map<int, std::vector<std::vector<long long> > > TensorI;
+  std::map<int, std::vector<std::vector<long long> > > TensorJ;
 
   // Pointers needed to extract V, I, J.
   int param_id;
@@ -36,12 +36,12 @@ public:
   void init_id(int new_param_id, int param_size) {
     assert(TensorV.count(new_param_id) == 0);
     std::vector<std::vector<double> > vecV(param_size);
-    std::vector<std::vector<int> > vecI(param_size);
-    std::vector<std::vector<int> > vecJ(param_size);
+    std::vector<std::vector<long long> > vecI(param_size);
+    std::vector<std::vector<long long> > vecJ(param_size);
     for (int i = 0; i < param_size; ++i) {
       std::vector<double> elemV;
-      std::vector<int> elemI;
-      std::vector<int> elemJ;
+      std::vector<long long> elemI;
+      std::vector<long long> elemJ;
       vecV.push_back(elemV);
       vecI.push_back(elemI);
       vecJ.push_back(elemJ);
@@ -87,8 +87,8 @@ public:
   /**
    * Returns the row index vector I as a contiguous 1D numpy array.
    */
-  void getI(int *values, int num_values) {
-    std::vector<int> I = TensorI[param_id][vec_idx];
+  void getI(long long *values, int num_values) {
+    std::vector<long long> I = TensorI[param_id][vec_idx];
     for (int i = 0; i < num_values; i++) {
       values[i] = I[i];
     }
@@ -97,8 +97,8 @@ public:
   /**
    * Returns the column index vector J as a contiguous 1D numpy array.
    */
-  void getJ(int *values, int num_values) {
-    std::vector<int> J = TensorJ[param_id][vec_idx];
+  void getJ(long long *values, int num_values) {
+    std::vector<long long> J = TensorJ[param_id][vec_idx];
     for (int i = 0; i < num_values; i++) {
       values[i] = J[i];
     }
